src/python_read_system/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
Django==4.0.6
django-ztask==0.1.5
djangorestframework==3.13.1
newrelic==7.16.0.178
protobuf==3.20.0
protobuf-to-dict==0.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pulsar-client==2.10.1
python-decouple==3.1
pytz==2022.1
pyzmq==23.2.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.11
zmq==0.0.0